6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float z;
 
-printf("size of a char: %lu byte(s)\n", sizeof(a));
-printf("size of an int: %lu byte(s)\n", sizeof(b));
-printf("size of a long int: %lu byte(s)\n", (sizeof(c));
-printf("size of a long long int: %lu byte(s)\n", sizeof(d));
-printf("size of a float: %lu byte(s)\n", sizeof(z));
+printf("size of a char: %lu byte(s)\n", (unsigned int) sizeof(a));
+printf("size of an int: %lu byte(s)\n", (unsigned int) sizeof(b));
+printf("size of a long int: %lu byte(s)\n", (unsigned int) sizeof(c);
+printf("size of a long long int: %lu byte(s)\n", (unsigned int) sizeof(d));
+printf("size of a float: %lu byte(s)\n", (unsigned int) sizeof(z));
 return (0);
 }
